requirements.txt
ultralytics>=8.0.0
torch>=1.12.0
torchvision>=0.13.0
opencv-python>=4.6.0
numpy>=1.22.0
Pillow>=9.2.0
requests>=2.28.1
pyyaml>=6.0
pycryptodome>=3.14.0
albumentations>=1.0.0
tqdm>=4.62.0
transformers>=4.40.0
datasets>=2.0.0
easyocr
accelerate>=0.26.0
paddleocr==2.8.0
lmdb
rapidfuzz
